requirements.txt
amqp==5.0.7
asgiref==3.4.1
billiard==3.6.4.0

certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.10
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-filter==21.1
django-import-export==2.7.0
django-timezone-field==4.2.1
djangorestframework==3.13.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwt==1.3.1

MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
pycparser==2.21
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
